% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp.R
\name{gsp}
\alias{gsp}
\title{Spatial mark variogram function}
\usage{
gsp(xyt, ds, ks = "epanech", hs)
}
\arguments{
\item{xyt}{Spatial coordinates and times \eqn{(x,y,t)} of the point pattern.}

\item{ds}{A vector of distances \code{u} at which \eqn{\gamma_{sp}(u)} is computed.}

\item{ks}{A kernel function for the spatial distances. The default is the \code{"box"} kernel. It can also be \code{"epanech"} for the Epanechnikov kernel, or \code{"biweight"}.}

\item{hs}{A bandwidth of the kernel function \code{ks}.}
}
\value{
A list containing:
\itemize{
  \item \code{gspke}: A vector containing the values of \eqn{\widehat{\gamma}_{sp}(r)} estimated.
  \item \code{ds}: If \code{ds} is missing, a vector of distances \code{u} at which \eqn{\gamma_{sp}(u)} is computed under the restriction \eqn{0<\epsilon<r}.
  \item \code{kernel}: A vector of names and bandwidth of the spatial kernel.
  }
}
\description{
Computes an estimator of the spatial mark variogram function.
}
\details{
\eqn{N = {[x_{1},t_{1}],[x_{2},t_{2}],...,[x_{n},t_{n}]}} with \eqn{x_{i}} in \eqn{W} subset of \eqn{R^2} and \eqn{t_{i}} in \eqn{T} subset of \eqn{R}, the spatial mark variogram \eqn{\gamma_{sp}(r)} considers the times \eqn{t_{i}} in the points \eqn{[x_{i},t_{i}]} as marks of the locations \eqn{x_{i}}. This approach makes sense if, as in statistical applications, the \eqn{t_{i}} come from some time interval \eqn{T}. The function \eqn{\gamma_{sp}(r)} is given by the mean of the half-squared differences of the times of points with spatial distance \eqn{r}, and the estimator is given by \deqn{\widehat{\gamma}_{sp}(r) = sum_{i = 1,...,n} sum_{j = 1,...,n; j! = i}(1/2 (t_{1}-t_{2})^2 \kappa_{\epsilon}(||x_{i}-x_{j}||-r)  / sum_{i = 1,...,n} sum_{j = 1,...,n; j != i} (\kappa_{\epsilon} (||x_{i}-x_{j}||-r)),} where \eqn{\kappa} is a one-dimensional kernel function with bandwidth \eqn{\epsilon}.
}
\examples{
## Not run:
#################

# Realisations of the homogeneous Poisson processes
hpp <- rpp(lambda = 100, replace = FALSE)$xyt

# R plot
plot(hpp)

# This function provides an kernel estimator of the spatial mark variogram
out <- gsp(hpp,hs=0.1/sqrt(100))

# R plot - Spatial mark variogram
par(mfrow=c(1,1))
gsp_range <- range(out$gspke,1/12)
plot(out$ds,out$gspke,type="l",ylim=gsp_range,xlab="r = distance",
ylab=expression(hat(gamma)[sp](r)),main="Spatial mark variogram")
lines(out$ds,rep(1/12,length(out$ds)),type="l",col="red")

## End(Not run)
}
\author{
Francisco J. Rodriguez-Cortes <cortesf@uji.es> \url{https://fjrodriguezcortes.wordpress.com}
}
\references{
Baddeley, A. and Turner, J. (2005). \code{spatstat}: An R Package for Analyzing Spatial Point Pattens. Journal of Statistical Software 12, 1-42.

Chiu, S. N., Stoyan, D., Kendall, W. S., and Mecke, J. (2013). Stochastic Geometry and its Applications. John Wiley & Sons.

Gabriel, E., Rowlingson, B., Diggle P J. (2013) \code{stpp}: an R package for plotting, simulating and analyzing Spatio-Temporal Point Patterns. Journal of Statistical Software 53, 1-29.

Illian, J B., Penttinen, A., Stoyan, H. and Stoyan, D. (2008). Statistical Analysis and Modelling of Spatial Point Patterns. John Wiley and Sons, London.
}

